Use cgo for critical operations
When a call to SetNS has been made, the Go runtime can
decide to create a new thread - for instance when a syscall
is done. This new thread inherits the namespace of our
first thread. This can lead to bugs that are hard to debug.

This commit changes some of the calls from pure Go -
that could trigger the Go runtime - to CGo calls.